requirements.txt
bcrypt==5.0.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
idna==3.11
invoke==2.2.1
lxml==6.0.2
ncclient==0.7.0
paramiko==4.0.0
pycparser==2.23
PyNaCl==1.6.0
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.5.0
xmltodict==1.0.2
